[fix] fixing includes at makefile and casting things

+++ src/idt/idt.c
@@ -1,11 +1,12 @@
 #include "idt.h"
 #include "config.h"
+#include "kernel.h"
 #include "memory/memory.h"
 
 struct idt_desc idt_descriptors[SOS_TOTAL_INTERRUPTS];
 struct idtr_desc idtr_descriptor;
 
-extern void idt_load(struct idtr_desc    *ptr);
+extern void idt_load(struct idtr_desc *ptr);
 
 void idt_zero()
 {
@@ -26,7 +27,7 @@ void idt_init()
 {
     memset(idt_descriptors, 0, sizeof(idt_descriptors));
     idtr_descriptor.limit = sizeof(idt_descriptors) - 1;
-    idtr_descriptor.base = idt_descriptors;
+    idtr_descriptor.base = (uint32_t)idt_descriptors;
     idt_set(0, idt_zero);
 
     // laod the idt
